requirements.txt
# requirements.txt

pyodbc
python-dotenv
tabulate  # Optional, but recommended for nicely formatted console output